Modify webclient code into portal plugin
This is a fairly large change that was actually done previously over
many hours of testing and untracked changes before it was actually
working. The repository was created in a specific sequence in order to
clarify exactly what has been changed here from Evennia's original
webclient code that I started with. Some changes may need to be
undone or fixed in some way, but it seems to work pretty much as intended.

Notable changes:
- Add code for utilizing as a portal service plugin
- Remove everything related to csessionid
- Remove html pre-processing