Change inheritance of AlpacasCharacter
Simple change for AlpacasCharacter to set up dual inheritance from
both AlpacasObject and DefaultObject. This allows characters' presence
in rendered scenes to utilize the same code as any other object.

While this works to a certain degree, AlpacasObject is currently
hardcoded to assign everything the "obj_type" key of "thing" in
the modified move_to() class method when a character moves around.